k8s: Separate out CiliumNetworkPolicy CRD type
This commit separates out CiliumNetworkPolicy from v2/types.go to
prepare for use with the controller-gen tool for CRD generation.
